chore: move ws tests under "socket.io"
